+++ CHANGELOG.md
@@ -687,3 +687,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Much more options for how events spawn (Alchyr)
 
 #### dev ####
+* Color tiny cards in Run History for modded cards (kiooeht)